requirements.txt
# WorldOfTaxanomy dependencies

# Phase 1: Data Model & Ingestion
openpyxl>=3.1.0
xlrd>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Phase 2: REST API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Phase 3: MCP Server
# No external dependency — protocol implemented directly (Python 3.9 compatible)

# Phase 4: Web Frontend
jinja2>=3.1.0

# Phase 5: Auth & Rate Limiting
bcrypt>=4.0.0
PyJWT>=2.8.0
slowapi>=0.1.9
